model/requirements.txt
ultralytics==8.1.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
pyyaml==6.0.1




